# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_rekall_backend_internal_interfaces_http_dto.AccessTokenResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.TokenObj'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.ChangePasswordRequest:
    properties:
      current_password:
        example: OldP@ssw0rd123
        type: string
      new_password:
        example: NewP@ssw0rd123
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.ErrorBody:
    properties:
      code:
        example: CALL_NOT_FOUND
        type: string
      details: {}
      message:
        example: call with the given ID was not found
        type: string
    type: object
  github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorBody'
      success:
        example: false
        type: boolean
    type: object
  github_com_rekall_backend_internal_interfaces_http_dto.ForgotPasswordRequest:
    properties:
      email:
        example: alice@example.com
        type: string
    required:
    - email
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.LoginRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      password:
        example: P@ssw0rd123
        type: string
    required:
    - email
    - password
    type: object
  github_com_rekall_backend_internal_interfaces_http_dto.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIwMDAwMDAwMC0wMDAwLTAwMDAtMDAwMC0wMDAwMDAwMDAwMDEifQ.signature
        type: string
      user:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UserResponse'
    type: object
  github_com_rekall_backend_internal_interfaces_http_dto.LoginResponseEnvelope:
    properties:
      data:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.LoginResponse'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.MessageResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MsgObj'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.MsgObj:
    properties:
      message:
        example: operation completed successfully
        type: string
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.RegisterRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      full_name:
        example: Alice Nguyen
        type: string
      password:
        example: P@ssw0rd123
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  github_com_rekall_backend_internal_interfaces_http_dto.ResendVerificationRequest:
    properties:
      email:
        example: alice@example.com
        type: string
    required:
    - email
    type: object
  github_com_rekall_backend_internal_interfaces_http_dto.ResetPasswordRequest:
    properties:
      password:
        example: NewP@ssw0rd123
        type: string
      token:
        example: a3f2c1d4e5b6...
        type: string
    required:
    - password
    - token
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.UpdateMeRequest:
    properties:
      full_name:
        example: Alice Nguyen
        maxLength: 100
        minLength: 1
        type: string
    required:
    - full_name
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.UserResponseEnvelope:
    properties:
      data:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UserResponse'
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: engineering@rekall.io
    name: Rekall Engineering
  description: |-
    Rekall is a call intelligence platform. Record, transcribe, and analyse voice and video calls.
    This API powers the Rekall web application and is available for direct integration.

    ## Authentication
    Protected endpoints require a JWT access token obtained from `POST /api/v1/auth/login`.
    Pass the token in the `Authorization` header as `Bearer <token>`.
    The refresh token is transported exclusively via an `HttpOnly` cookie set by the server.

    ## Response Envelope
    All responses are wrapped in a standard envelope:
    - Success: `{ "success": true, "data": <T>, "meta": <pagination|null> }`
    - Error:   `{ "success": false, "error": { "code": "ERROR_CODE", "message": "Human readable message", "details": <null|object> } }`

    ## Error Codes
    Every error response includes a machine-readable `code` field. Common codes:
    `INVALID_CREDENTIALS`, `EMAIL_NOT_VERIFIED`, `EMAIL_ALREADY_REGISTERED`,
    `INVALID_OR_EXPIRED_TOKEN`, `INVALID_REFRESH_TOKEN`, `MISSING_TOKEN`, `INVALID_TOKEN`,
    `CALL_NOT_FOUND`, `USER_NOT_FOUND`, `SLUG_ALREADY_TAKEN`, `ALREADY_A_MEMBER`,
    `CANNOT_REMOVE_OWNER`, `INVITATION_EMAIL_MISMATCH`, `DEPARTMENT_NOT_FOUND`, `FORBIDDEN`
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Rekall API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password. On success returns
        a short-lived JWT **access_token** (15 min) and sets an `HttpOnly Secure SameSite=Strict`
        **refresh_token** cookie (7 days).
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authenticated successfully
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.LoginResponseEnvelope'
        "401":
          description: Invalid credentials (INVALID_CREDENTIALS)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "403":
          description: Email address not yet verified (EMAIL_NOT_VERIFIED)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      summary: Sign in
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Revokes the refresh token stored in the `HttpOnly` cookie and clears
        the cookie. Access tokens are stateless and expire naturally — no server-side
        invalidation is needed.
      produces:
      - application/json
      responses:
        "204":
          description: Signed out — refresh token revoked and cookie cleared
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      summary: Sign out
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      description: 'Returns the full profile of the authenticated user. Requires a
        valid JWT access token in the `Authorization: Bearer <token>` header.'
      produces:
      - application/json
      responses:
        "200":
          description: Authenticated user profile
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UserResponseEnvelope'
        "401":
          description: Missing or invalid token (MISSING_TOKEN / INVALID_TOKEN)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "403":
          description: Email address not verified (EMAIL_NOT_VERIFIED)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's profile. Only the full_name field
        is editable via this endpoint; email, role, and verified status require different
        flows. Silently ignores any additional fields in the request body.
      parameters:
      - description: Profile fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UpdateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UserResponseEnvelope'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - Auth
  /api/v1/auth/password/change:
    post:
      consumes:
      - application/json
      description: Verifies the authenticated user's current password and installs
        a new one. Revokes every OTHER refresh token for the user — forcing other
        devices to re-authenticate — while issuing a freshly rotated refresh cookie
        for the current browser so this session stays signed in. Password must be
        8+ characters with at least one letter and one digit.
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated; other sessions signed out
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MessageResponse'
        "400":
          description: Current password is incorrect (INVALID_CURRENT_PASSWORD)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "422":
          description: Validation error — new password policy not met
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Auth
  /api/v1/auth/password/forgot:
    post:
      consumes:
      - application/json
      description: Sends a password reset link to the given email address. **Always
        returns 200 regardless of whether the email is registered** to prevent user
        enumeration. The reset link is valid for 1 hour.
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reset email dispatched (if account exists)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MessageResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      summary: Request password reset
      tags:
      - Auth
  /api/v1/auth/password/reset:
    post:
      consumes:
      - application/json
      description: Sets a new password using a valid, unexpired password reset token.
        On success all active refresh tokens for the user are revoked, forcing re-login
        on all devices. Password must be 8+ characters with at least one letter and
        one digit.
      parameters:
      - description: Reset token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated — all sessions terminated
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MessageResponse'
        "400":
          description: Token invalid, expired, or already used (INVALID_OR_EXPIRED_TOKEN)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "422":
          description: Validation error — password complexity requirements not met
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      summary: Reset password
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      description: Exchanges the `refresh_token` cookie for a new JWT access token.
        The refresh token is **rotated** — the old one is revoked and a new cookie
        is issued atomically.
      produces:
      - application/json
      responses:
        "200":
          description: New access token issued
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.AccessTokenResponse'
        "401":
          description: Refresh token invalid, expired, or revoked (INVALID_REFRESH_TOKEN)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      summary: Refresh access token
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email and password. The user receives
        a verification email and **cannot sign in until their email is confirmed**.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Account created — check email for a verification link
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UserResponseEnvelope'
        "409":
          description: Email already registered (EMAIL_ALREADY_REGISTERED)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "422":
          description: Validation error — see details for field-level errors
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /api/v1/auth/verify:
    get:
      description: Marks the user's email as verified using the single-use token delivered
        in the verification email. The token is valid for 24 hours.
      parameters:
      - description: Raw verification token from the email link
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MessageResponse'
        "400":
          description: Token missing, invalid, expired, or already used (INVALID_OR_EXPIRED_TOKEN)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      summary: Verify email address
      tags:
      - Auth
  /api/v1/auth/verify/resend:
    post:
      consumes:
      - application/json
      description: Generates a new verification token, invalidates any prior pending
        token, and resends the verification email to the given address.
      parameters:
      - description: Email address to resend verification to
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verification email sent (if the address is registered and unverified)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MessageResponse'
        "400":
          description: Email is already verified (EMAIL_ALREADY_VERIFIED)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      summary: Resend verification email
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT access token as: **Bearer &lt;token&gt;**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"